feat: update migration

+++ packages/indexer/src/migrations/1682624868123_tokens-supply.sql
@@ -1,7 +1,7 @@
 -- Up Migration
 
-ALTER TABLE "tokens" ADD COLUMN "supply" INT;
-ALTER TABLE "tokens" ADD COLUMN "remaining_supply" INT;
+ALTER TABLE "tokens" ADD COLUMN "supply" NUMERIC(78, 0);
+ALTER TABLE "tokens" ADD COLUMN "remaining_supply" NUMERIC(78, 0);
 
 -- todo run this manually and uncomment after deployment
 --CREATE INDEX "tokens_contract_remaining_supply_collection_id_index"
